+++ mocket/core/socket.py
@@ -13,13 +13,13 @@ from typing_extensions import Self
 from mocket.core.io import MocketSocketIO
 from mocket.core.mocket import Mocket
 from mocket.core.mode import MocketMode
-from mocket.entry import MocketEntry
-from mocket.types import (
+from mocket.core.types import (
     Address,
     ReadableBuffer,
     WriteableBuffer,
     _RetAddress,
 )
+from mocket.entry import MocketEntry
 
 true_gethostbyname = socket.gethostbyname
 true_socket = socket.socket
